backend/pyproject.toml
[project]
name = "sniffleads"
version = "0.1.0"
description = "Lead generation crawler and API"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14,<4.0",
    "celery[redis]>=5.3,<6.0",
    "redis>=5.0,<6.0",
    "httpx>=0.27,<1.0",
    "playwright>=1.40,<2.0",
    "psycopg[binary]>=3.1,<4.0",
    "python-decouple>=3.8,<4.0",
    "gunicorn>=21.0,<23.0",
    "beautifulsoup4>=4.12,<5.0",
    "lxml>=5.0,<6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.7,<5.0",
    "pytest-celery>=0.0.0",
    "ipython>=8.0,<9.0",
    "ruff>=0.1,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "sniffleads"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sniffleads.settings.local"
python_files = ["test_*.py"]